Add rectangle animations
